Fixed circular dependencies reported when packaging
