+++ src/shared/c8ypact/c8ypact.ts
@@ -5,21 +5,9 @@ import lodash1 from "lodash";
 import * as lodash2 from "lodash";
 const _ = lodash1 || lodash2;
 
-import {
-  C8yPactPreprocessor,
-  C8yPactPreprocessorOptions,
-} from "./preprocessor";
-import { C8yClient, C8yClientOptions } from "../c8yclient";
-import { C8ySchemaGenerator } from "./schema";
-import { removeBaseUrlFromRequestUrl } from "./url";
-import {
-  C8yAuthOptions,
-  C8yPactAuthObject,
-  toPactAuthObject,
-  isAuthOptions,
-  isPactAuthObject,
-} from "../auth";
-import { C8yDefaultPact } from "./c8ydefaultpact";
+import { C8yPactPreprocessorOptions } from "./preprocessor";
+import { C8yClientOptions } from "../c8yclient";
+import { C8yPactAuthObject } from "../auth";
 
 export const C8yPactModeValues = [
   "record",
@@ -288,111 +276,6 @@ export interface C8yPactRecord {
   date(): Date | null;
 }
 
-/**
- * Default implementation of C8yPactRecord. Use C8yDefaultPactRecord.from to create
- * a C8yPactRecord from a Cypress.Response object or an C8yPactRecord object.
- */
-export class C8yDefaultPactRecord implements C8yPactRecord {
-  request: C8yPactRequest;
-  response: C8yPactResponse<any>;
-  options?: C8yClientOptions;
-  auth?: C8yPactAuthObject;
-  createdObject?: string;
-  modifiedResponse?: C8yPactResponse<any>;
-
-  constructor(
-    request: C8yPactRequest,
-    response: C8yPactResponse<any>,
-    options?: C8yClientOptions,
-    auth?: C8yPactAuthObject,
-    createdObject?: string,
-    modifiedResponse?: C8yPactResponse<any>
-  ) {
-    this.request = request;
-    this.response = response;
-
-    if (options) this.options = options;
-    if (auth) this.auth = auth;
-    if (createdObject) this.createdObject = createdObject;
-    if (modifiedResponse) this.modifiedResponse = modifiedResponse;
-
-    if (request?.method?.toLowerCase() === "post") {
-      const newId = response.body?.id;
-      if (newId) {
-        this.createdObject = newId;
-      }
-    }
-  }
-
-  /**
-   * Creates a C8yPactRecord from a Cypress.Response or an C8yPactRecord object.
-   * @param obj The Cypress.Response<any> or C8yPactRecord object.
-   * @param client The C8yClient for options and auth information.
-   */
-  static from(
-    obj: Cypress.Response<any> | C8yPactRecord | Partial<Cypress.Response<any>>,
-    auth?: C8yAuthOptions,
-    client?: C8yClient
-  ): C8yPactRecord {
-    // if (obj == null) return obj;
-    if ("request" in obj && "response" in obj) {
-      return new C8yDefaultPactRecord(
-        _.get(obj, "request"),
-        _.get(obj, "response"),
-        _.get(obj, "options") || {},
-        _.get(obj, "auth"),
-        _.get(obj, "createdObject"),
-        _.get(obj, "modifiedResponse")
-      );
-    }
-
-    const r = _.cloneDeep(obj);
-    return new C8yDefaultPactRecord(
-      toPactRequest(r) || {},
-      toPactResponse(r) || {},
-      client?._options,
-      isAuthOptions(auth) || isPactAuthObject(auth)
-        ? toPactAuthObject(auth)
-        : client?._auth
-        ? toPactAuthObject(client?._auth)
-        : undefined
-    );
-  }
-
-  /**
-   * Returns the date of the response.
-   */
-  date(): Date | null {
-    const date = _.get(this.response, "headers.date");
-    if ((date && _.isString(date)) || _.isNumber(date) || _.isDate(date)) {
-      return new Date(date);
-    }
-    return null;
-  }
-
-  /**
-   * Converts the C8yPactRecord to a Cypress.Response object.
-   */
-  toCypressResponse<T>(): Cypress.Response<T> {
-    const result = _.cloneDeep(this.response);
-    _.extend(result, {
-      ...(result.status && {
-        isOkStatusCode: result.status > 199 && result.status < 300,
-      }),
-      ...(this.request.headers && {
-        requestHeaders: Object.fromEntries(
-          Object.entries(this.request.headers || [])
-        ),
-      }),
-      ...(this.request.url && { url: this.request.url }),
-      ...(result.allRequestResponses && { allRequestResponses: [] }),
-      ...(this.request.body && { requestBody: this.request.body }),
-      method: this.request.method || this.response.method || "GET",
-    });
-    return result as Cypress.Response<T>;
-  }
-}
-
 export function isValidPactId(value: string): boolean {
   if (value == null || value.length > 1000 || !_.isString(value)) return false;
   const validPactIdRegex = /^[a-zA-Z0-9_-]+(__[a-zA-Z0-9_-]+)*$/;
@@ -567,108 +450,6 @@ export function toPactResponse<T>(
 }
 
 export type C8yPactSaveKeys = "id" | "info" | "records";
-
-export async function toPactSerializableObject(
-  response: Partial<Cypress.Response<any>>,
-  info: C8yPactInfo,
-  options: {
-    preprocessor?: C8yPactPreprocessor;
-    client?: C8yClient;
-    modifiedResponse?: Cypress.Response<any>;
-    schemaGenerator?: C8ySchemaGenerator;
-    loggedInUser?: string;
-    loggedInUserAlias?: string;
-    authType?: string;
-  } = {}
-): Promise<Pick<C8yPact, C8yPactSaveKeys>> {
-  const recordOptions = {
-    loggedInUser: options?.loggedInUser,
-    loggedInUserAlias: options?.loggedInUserAlias,
-    authType: options?.authType,
-  };
-  const record = createPactRecord(response, options?.client, recordOptions);
-  removeBaseUrlFromRequestUrl(record, info.baseUrl);
-  const pact = new C8yDefaultPact([record], info, info.id);
-
-  if (
-    options?.modifiedResponse &&
-    isCypressResponse(options?.modifiedResponse)
-  ) {
-    const modifiedPactRecord = createPactRecord(
-      options.modifiedResponse,
-      options?.client,
-      recordOptions
-    );
-    pact.records[pact.records.length - 1].modifiedResponse =
-      modifiedPactRecord.response;
-  }
-
-  options?.preprocessor?.apply(pact);
-
-  const keysToSave: C8yPactSaveKeys[] = ["id", "info", "records"];
-  try {
-    await Promise.all(
-      pact.records
-        .filter(
-          (record_1) =>
-            record_1.response.body &&
-            !record_1.response.$body &&
-            _.isObjectLike(record_1.response.body)
-        )
-        .map((record_2) =>
-          options?.schemaGenerator
-            ?.generate(record_2.response.body, { name: "body" })
-            .then((schema) => {
-              record_2.response.$body = schema;
-              return record_2;
-            })
-        )
-    );
-    return { ..._.pick(pact, keysToSave) };
-  } catch (error) {
-    console.error(error);
-    return { ..._.pick(pact, keysToSave) };
-  }
-}
-
-export function createPactRecord(
-  response: Partial<Cypress.Response<any>>,
-  client?: C8yClient,
-  options: {
-    loggedInUser?: string;
-    loggedInUserAlias?: string;
-    authType?: string;
-  } = {}
-): C8yPactRecord {
-  let auth: C8yAuthOptions | undefined = undefined;
-  const envUser = options.loggedInUser;
-  const envAlias = options.loggedInUserAlias;
-  const envType = options.authType;
-  const envAuth = {
-    ...(envUser && { user: envUser }),
-    ...(envAlias && { userAlias: envAlias }),
-    ...(envAlias && { type: envType ?? "CookieAuth" }),
-  };
-
-  if (client?._auth) {
-    // do not pick the password. passwords must not be stored in the pact.
-    auth = _.defaultsDeep(client._auth, envAuth);
-    if (client._auth.constructor != null) {
-      if (!auth) {
-        auth = { type: client._auth.constructor.name };
-      } else {
-        auth.type = client._auth.constructor.name;
-      }
-    }
-  }
-  if (!auth && (envUser || envAlias)) {
-    auth = envAuth;
-  }
-
-  // only store properties that need to be exposed. do not store password.
-  auth = auth ? toPactAuthObject(auth) : auth;
-  return C8yDefaultPactRecord.from(response, auth, client);
-}
 
 /**
  * Returns the value of the environment variable with the given name. The function
