+++ src/shared/c8ypact/index.ts
@@ -1,5 +1,6 @@
 export * from "./c8ypact";
 export * from "./c8ydefaultpact";
+export * from "./c8ydefaultpactrecord";
 export * from "./schema";
 export * from "./matcher";
 export * from "./preprocessor";
